2004-03-21 17:00 UTC+0100 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * xharbour/config/dos/watcom.cf     * divided compilation flags for easier manipulating in special builds
  * xharbour/config/linux/gcc.cf
    + -W added to report some warnings

  * xharbour/source/rtl/cdpapi.c
    * utf8tou16() temporary disabled to avoid warning

  * xharbour/source/rtl/gtcgi/gtcgi.c
  * xharbour/source/rtl/gtsln/gtsln.c
  * xharbour/source/rtl/gtsln/kbsln.c
  * xharbour/source/rtl/gtsln/keytrans.c
  * xharbour/source/rtl/strmatch.c
  * xharbour/source/vm/classes.c
  * xharbour/source/vm/garbage.c
  * xharbour/source/vm/hvm.c
  * xharbour/source/rdd/dbcmd.c
  * xharbour/source/rdd/dbf1.c
  * xharbour/source/rdd/workarea.c
  * xharbour/contrib/rdd_ads/adsfunc.c
  * xharbour/contrib/rdd_ads/ads1.c

    * warning fixes

  * xharbour/source/rtl/filesys.c
    * fixed some warnings
    ! fixed hb_fsOpen for OpenWatcom. It doesn't respect shared flags
      at all. I change the order of #if code and some definition.
      Most of them was redundant and cause that is was hard to understand
      what each compiler should do - side effect of such situation cause
      that some sompilers may not use shared flags at all (f.e. OpenWatcom)
      I hope I don't break anything but this should be extensively tested
      on all platforms.

+++ xharbour/source/rtl/gtsln/kbsln.c
@@ -1,5 +1,5 @@
 /*
- * $Id: kbsln.c,v 1.10 2003/10/19 18:03:29 lculik Exp $
+ * $Id: kbsln.c,v 1.11 2003/10/22 01:04:21 lculik Exp $
  */
 
 /*
@@ -335,7 +335,7 @@ int HB_GT_FUNC(gt_ReadKey( HB_inkey_enum eventmask ))
             return( 0 );
 #endif
     /* user AbortKey break */
-    if( ch == s_hb_gt_Abort_key )
+    if( (int) ch == s_hb_gt_Abort_key )
         return( HB_BREAK_FLAG );
 
     SLang_ungetkey( ch );
@@ -351,18 +351,18 @@ int HB_GT_FUNC(gt_ReadKey( HB_inkey_enum eventmask ))
     if( InDeadState )
     {
         InDeadState = FALSE;
-        if( ch == hb_DeadKey ) /* double press Dead key */
+        if( (int) ch == hb_DeadKey ) /* double press Dead key */
             return( ch );
         if( ch < 256 )  /* is this needed ??? */
         {
             int i;
             for( i=0; i < ( int ) s_convKDeadKeys[ 0 ]; i++ )
-                if( ( int ) s_convKDeadKeys[ 2 * i + 1 ] == ch )
+                if( ( int ) s_convKDeadKeys[ 2 * i + 1 ] == (int) ch )
                     return( ( int ) s_convKDeadKeys[ 2 * i + 2 ] );
         }
         return( 0 );
     }
-    else if( ch == hb_DeadKey )
+    else if( (int) ch == hb_DeadKey )
     {
         /* entering Dead key state */
         InDeadState = TRUE;
